requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cloudinary==1.31.0
dj-database-url==1.2.0
Django==4.1.6
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.3.0
